# Only allocate named logger if necessary in activator