<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

Allocating a named logger is somewhat expensive (~ 10 allocs) so it's worth avoiding it, especially in high-speed components like the activator.

This defers allocation of the respective named logger to error cases.

## Context handler benchmark

```
benchmark                                                         old ns/op     new ns/op     delta
BenchmarkContextHandler/context_handler_success-sequential-16     3258          924           -71.64%
BenchmarkContextHandler/context_handler_success-parallel-16       569           175           -69.24%
BenchmarkContextHandler/context_handler_failure-sequential-16     14659         14838         +1.22%
BenchmarkContextHandler/context_handler_failure-parallel-16       10980         11384         +3.68%

benchmark                                                         old allocs     new allocs     delta
BenchmarkContextHandler/context_handler_success-sequential-16     19             7              -63.16%
BenchmarkContextHandler/context_handler_success-parallel-16       19             7              -63.16%
BenchmarkContextHandler/context_handler_failure-sequential-16     45             38             -15.56%
BenchmarkContextHandler/context_handler_failure-parallel-16       45             38             -15.56%

benchmark                                                         old bytes     new bytes     delta
BenchmarkContextHandler/context_handler_success-sequential-16     2091          592           -71.69%
BenchmarkContextHandler/context_handler_success-parallel-16       2090          592           -71.67%
BenchmarkContextHandler/context_handler_failure-sequential-16     5086          4119          -19.01%
BenchmarkContextHandler/context_handler_failure-parallel-16       4882          3715          -23.90%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @julz @vagababov